Role Update for users

+++ src/main/java/com/api/spring/boot/funsho/api/entity/users.java
@@ -53,7 +53,7 @@ public class users {
     @Column(length = 500)
     String bio;
 
-    int role;
+    Long role;
     
 
     @OneToOne(
